requirements.txt
allure-pytest==2.8.16
allure-python-commons==2.8.16
asgiref==3.2.10
atomicwrites==1.4.0
attrs==19.3.0
colorama==0.4.3
jsonpath==0.82
more-itertools==8.4.0
mysqlclient==2.0.1
packaging==20.4
pluggy==0.13.1
py==1.9.0
PyMySQL==0.9.3
pyparsing==2.4.7
pytest==5.4.2
pytest-html==2.1.1
pytest-metadata==1.10.0
pytz==2020.1
PyYAML==5.3.1
selenium==3.141.0
six==1.15.0
sqlparse==0.3.1
wcwidth==0.2.5
